requirements.txt
datasets==3.5.0
lightning==2.5.1
transformers==4.51.0
accelerate==1.6.0
wandb==0.19.10
hydra-core==1.3.2
dvc==3.59.2
onnx==1.18.0
onnxruntime==1.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
cryptography==43.0.3
datasets==3.5.0
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
dpath==2.2.0
dulwich==0.22.8
dvc==3.59.2
dvc-data==3.16.10
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
flatten-dict==0.4.2
flufl.lock==8.2.0
frozenlist==1.6.0
fsspec==2024.12.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grandalf==0.8
grpcio==1.71.0
grpcio-status==1.71.0
gto==1.7.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
iterative-telemetry==0.0.10
Jinja2==3.1.6
jiter==0.10.0
kombu==5.5.3
lightning==2.5.1
lightning-utilities==0.14.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.6
oauth2client==4.1.3
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
openai==1.79.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==4.0.0
PyDrive2==1.21.3
pygit2==1.18.0
Pygments==2.19.1
pygtrie==2.5.0
pyOpenSSL==24.2.1
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.5.1.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
readmeai==0.6.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scmrepo==3.3.11
semver==3.0.4
sentry-sdk==2.28.0
setproctitle==1.3.6
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqltrie==0.11.2
starlette==0.46.2
structlog==24.4.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.4.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.7.0
torchmetrics==1.7.1
tqdm==4.67.1
transformers==4.51.0
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
voluptuous==0.15.2
wandb==0.19.10
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.0
zc.lockfile==3.0.post1